Refactor self-activation methods into separate views
