+++ client/react/app/components/self-activation/index.ts
@@ -2,10 +2,28 @@ import SelfActivationMethods from './SelfActivationMethods';
 import CognitiveDistortions from './CognitiveDistortions';
 import DailyImperatives from './DailyImperatives';
 import DailyGratitude from './DailyGratitude';
+import AntiProcrastinationSheet from './AntiProcrastinationSheet';
+import PleasurePredicting from './PleasurePredicting';
+import TripleColumnTechnique from './TripleColumnTechnique';
+import ButRebuttal from './ButRebuttal';
+import SelfEndorsement from './SelfEndorsement';
+import TicTocTechnique from './TicTocTechnique';
+import MotivationWithoutCoercion from './MotivationWithoutCoercion';
+import CannotLoseSystem from './CannotLoseSystem';
+import VisualizeSuccess from './VisualizeSuccess';
 
 export {
   SelfActivationMethods,
   CognitiveDistortions,
   DailyImperatives,
-  DailyGratitude
+  DailyGratitude,
+  AntiProcrastinationSheet,
+  PleasurePredicting,
+  TripleColumnTechnique,
+  ButRebuttal,
+  SelfEndorsement,
+  TicTocTechnique,
+  MotivationWithoutCoercion,
+  CannotLoseSystem,
+  VisualizeSuccess
 };
