+++ client/react/app/components/self-activation/methods.ts
@@ -2,72 +2,86 @@ const METHODS = [
   {
     symptoms: 'You feel disorganized. You have nothing to do. You get lonely and bored on weekends.',
     technique: 'Daily Activity Schedule',
-    purpose: 'Plan things one hour at a time and record the amount of master and pleasure. Virtually any activity will make you feel better than lying in bed and will undercut your sense of inadequacy.'
+    purpose: 'Plan things one hour at a time and record the amount of master and pleasure. Virtually any activity will make you feel better than lying in bed and will undercut your sense of inadequacy.',
+    href: '#'
   },
   {
     symptoms: 'You procrastinate because tasks seem too difficult and unrewarding.',
     technique: 'The Antiprocrastination Sheet',
-    purpose: 'You put your negative predictions to the test.'
+    purpose: 'You put your negative predictions to the test.',
+    href: '/anti-procrastination'
   },
   {
     symptoms: 'You feel overwhelmed by the urge to do nothing.',
     technique: 'Daily Record of Dysfunctional Thoughts',
-    purpose: 'You expose the illogical thoughts that paralyze you. You learn that motivation follows action, not vice versa.'
+    purpose: 'You expose the illogical thoughts that paralyze you. You learn that motivation follows action, not vice versa.',
+    href: '#'
   },
   {
     symptoms: 'You feel like there\'s no point in doing anything when you\'re alone.',
     technique: 'Pleasure-Predicting Sheet',
-    purpose: 'Schedule activities with the potential for personal growth or satisfaction, and predict how rewarding they will be. Compare the actual satisfaction you experience when you are alone and when you are with others.'
+    purpose: 'Schedule activities with the potential for personal growth or satisfaction, and predict how rewarding they will be. Compare the actual satisfaction you experience when you are alone and when you are with others.',
+    href: '/pleasure-predicting'
   },
   {
     symptoms: 'You give yourself excuses for avoiding things.',
     technique: 'But-Rebuttal',
-    purpose: 'You get off your "but" by combatting your "buts" with realistic rebuttals.'
+    purpose: 'You get off your "but" by combatting your "buts" with realistic rebuttals.',
+    href: '/but-rebuttal'
   },
   {
     symptoms: 'You have the idea that whatever you do isn\'t worth much.',
     technique: 'Self-Endorsement',
-    purpose: 'Write down the self-downing thoughts and talk back to them. Look for distorted thought patterns, such as "all-or-nothing thinking." Make a list of things you do accomplish each day.'
+    purpose: 'Write down the self-downing thoughts and talk back to them. Look for distorted thought patterns, such as "all-or-nothing thinking." Make a list of things you do accomplish each day.',
+    href: '/self-endorsement'
   },
   {
     symptoms: 'You think about a task in a self-defeating manner.',
     technique: 'TIC-TOC Technique',
-    purpose: 'You substitute task-oriented cognitions (TOCs) for task-interfering cognitions (TICs).'
+    purpose: 'You substitute task-oriented cognitions (TOCs) for task-interfering cognitions (TICs).',
+    href: '/tic-toc'
   },
   {
     symptoms: 'You feel overwhelmed by the magnitude of everything you have to do.',
     technique: 'Little Steps for Little Feet',
-    purpose: 'Break the task down into its tiny component parts, and do these one step at a time.'
+    purpose: 'Break the task down into its tiny component parts, and do these one step at a time.',
+    href: '#'
   },
   {
     symptoms: 'You feel guilty, oppressed, obliged, and duty-bound.',
     technique: 'Motivation Without Coercion',
-    purpose: 'You eliminate "shoulds," "musts," and "oughts," when you give yourself instructions. You list the advantages and disadvantages of any activity so you can begin to think in terms of what you want to do rather than what you must do.'
+    purpose: 'You eliminate "shoulds," "musts," and "oughts," when you give yourself instructions. You list the advantages and disadvantages of any activity so you can begin to think in terms of what you want to do rather than what you must do.',
+    href: '/motivation-without-coercion'
   },
   {
     symptoms: 'Someone else nags and harangues you. You feel pressured and resentful, so you refuse to do anything at all.',
     technique: 'Disarming Technique',
-    purpose: 'You assertively agree with them and remind them that you are capable of doing your own thinking.'
+    purpose: 'You assertively agree with them and remind them that you are capable of doing your own thinking.',
+    href: '#'
   },
   {
     symptoms: 'You have difficult modifying a habit such as smoking.',
     technique: 'Visualize Success',
-    purpose: 'You make a list of the positive benefits of having changed the habit. You visualize these after inducing a state of deep relaxation.'
+    purpose: 'You make a list of the positive benefits of having changed the habit. You visualize these after inducing a state of deep relaxation.',
+    href: '/visualize-success'
   },
   {
     symptoms: 'You feel unable to do anything on your own initiative because you see yourself as "a procrastinator."',
     technique: 'Count What Counts',
-    purpose: 'You count the things you do each day on your own initiative, using a wrist counter. This helps you overcome your bad habit of constantly dwelling on your inadequacies.'
+    purpose: 'You count the things you do each day on your own initiative, using a wrist counter. This helps you overcome your bad habit of constantly dwelling on your inadequacies.',
+    href: '#'
   },
   {
     symptoms: 'You feel inadequate and incompetent because you say, "I can\'t."',
     technique: 'Test Your Can\'ts',
-    purpose: 'You set up an experiment in which you challenge and disprove your negative predictions.'
+    purpose: 'You set up an experiment in which you challenge and disprove your negative predictions.',
+    href: '#'
   },
   {
     symptoms: 'You are afraid to fail, so you risk nothing.',
     technique: '"Can\'t Lose" System',
-    purpose: 'Write down any negative consequences of failure and develop a coping strategy ahead of time.'
+    purpose: 'Write down any negative consequences of failure and develop a coping strategy ahead of time.',
+    href: '/cannot-lose'
   }
 ];
